Moving Event calls to Event enums (task #5394)
